+++ jannovar-htsjdk/src/test/java/de/charite/compbio/jannovar/mendel/filter/GeneWiseInheritanceProcessorOneAffectedTest.java
@@ -72,7 +72,7 @@ public class GeneWiseInheritanceProcessorOneAffectedTest {
 
 		ArrayList<VariantContext> result = new ArrayList<>();
 		try (GeneWiseMendelianAnnotationProcessor proc = new GeneWiseMendelianAnnotationProcessor(trio, jannovarDB,
-				vc -> result.add(vc))) {
+				reader.getFileHeader().getSequenceDictionary(), vc -> result.add(vc))) {
 			for (VariantContext vc : variants)
 				proc.put(vc);
 		}
@@ -114,7 +114,7 @@ public class GeneWiseInheritanceProcessorOneAffectedTest {
 
 		ArrayList<VariantContext> result = new ArrayList<>();
 		try (GeneWiseMendelianAnnotationProcessor proc = new GeneWiseMendelianAnnotationProcessor(trio, jannovarDB,
-				vc -> result.add(vc))) {
+				reader.getFileHeader().getSequenceDictionary(), vc -> result.add(vc))) {
 			for (VariantContext vc : variants)
 				proc.put(vc);
 		}